Rename function getStackConfig to readStackConfig
