+++ src/cdk/create-stack.ts
@@ -1,5 +1,5 @@
 import {App, Stack} from 'aws-cdk-lib';
-import type {StackConfig} from '../get-stack-config';
+import type {StackConfig} from '../read-stack-config';
 import {getDomainName} from '../utils/get-domain-name';
 import {getStackName} from '../utils/get-stack-name';
